pyproject.toml
[tool.poetry]
name = "your-assistant"
version = "0.0.1"
description = "Your personal assistant that can use your own data stored locally."
authors = ["Yexi Jiang <yjiang@Yexis-MacBook-Air.local>"]
readme = "README.md"
packages = [{include = "your_assistant"}]

[tool.poetry.dependencies]
python = "^3.9"
langchain = "^0.0.123"
openai = "^0.27.2"
google-api-python-client = "^2.82.0"
revchatgpt = "^4.0.9"
pytest = "^7.2.2"
python-dotenv = "^1.0.0"
flask = "^2.2.3"
discord = "^2.2.2"
discord-py = "^2.2.2"
certifi = "^2022.12.7"
dpytest = "^0.6.3"
pytest-asyncio = "^0.21.0"
pytest-mock = "^3.10.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
flake8 = "^6.0.0"
black = "^23.1.0"
mypy = "^1.1.1"
Sphinx = "^6.1.3"
pre-commit = "^3.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
